+++ mercury/mercury-library/agent-cli-didcommx/src/main/scala/io/iohk/atala/AgentClientAlice.scala
@@ -5,9 +5,7 @@ import zio._
 import zhttp.service.{ChannelFactory, Client, EventLoopGroup}
 import zhttp.http.{Method, Headers}
 
-import io.iohk.atala.mercury.Agent
-import io.iohk.atala.mercury.AgentService
-import io.iohk.atala.mercury.MediaTypes
+import io.iohk.atala.mercury._
 import io.iohk.atala.mercury.{given}
 import io.circe.Printer
 import io.circe.syntax._
@@ -21,8 +19,8 @@ import io.circe._, io.circe.parser._
 
 @main def AgentClientAlice() = {
 
-  val env = ChannelFactory.auto ++ EventLoopGroup.auto()
-  val app = AgentPrograms.pickupMessageProgram.provide(env, AgentService.alice)
+  val app = AgentPrograms.pickupMessageProgram
+    .provide(AgentService.alice, HttpClientZhttp.layer)
 
   Unsafe.unsafe { implicit u => Runtime.default.unsafe.run(app).getOrThrowFiberFailure() }